// DXC_Frontend/src/api/models/iconGroupDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * DXC_Core.API
 * OpenAPI spec version: 1.0
 */

export interface IconGroupDto {
  publicId?: string
  iconCategoryPublicId: string
  /**
   * @minLength 1
   * @maxLength 100
   */
  name: string
  /**
   * @maxLength 500
   * @nullable
   */
  description?: string | null
  displayOrder?: number
  thuTu?: number
  isActive?: boolean
  createdAt?: string
  updatedAt?: string
  /**
   * @maxLength 500
   * @nullable
   */
  imageUrl?: string | null
  /** @nullable */
  imagePublicId?: string | null
  /** @nullable */
  iconCategoryName?: string | null
  totalIcons?: number
}